+++ Tests/GeometriaTests/3D/Protocols/Vector/Vector3TypeTests.swift
@@ -42,12 +42,6 @@ class Vector3TypeTests: XCTestCase {
         XCTAssertEqual(sut, .init(x: 0, y: 0, z: 3))
     }
     
-    func testTake() {
-        let sut = Vector3D(x: 1, y: 2, z: 3)
-        
-        XCTAssertEqual(sut.take.underlying, sut)
-    }
-    
     func testInitRepeating() {
         let sut = Vector3D(repeating: 1)